created basic new event manager
